fix(runtime): align nonterminal workflow query
